fix: critical setlist indexing bugs — saves, song switching, GitHub Pages 404
All song lookups used SONGS[currentIndex] instead of getSetlistSongs()[currentIndex],
causing edits to save under wrong song titles, blocking song switching after edits,
and losing all data on refresh. Also fixed: bottom bar handler order (exitEditMode
before changing currentIndex), GitHub Pages 404 from absolute paths in manifest/sw,
capo badge display, XSS in thumbnails, server stability (SSE disconnect handling,
poll timeout, jobId collisions), and upgraded to BTC-PL model for chord detection.

+++ server/analyze.py
@@ -395,7 +395,7 @@ if __name__ == '__main__':
         sys.exit(1)
 
     try:
-        hint = int(sys.argv[2]) if len(sys.argv) > 2 else 0
+        hint = float(sys.argv[2]) if len(sys.argv) > 2 else 0
         result = analyze(sys.argv[1], start_time_hint=hint)
         print(json.dumps(result))
     except Exception as e:
